make code organization changes as per suggestions from Shubham
