+++ core/data_compressor.py
@@ -1,14 +1,6 @@
 import numpy as np
-from core.data_stream import BitsDataStream
-from core.data_transformer import (
-    BitsToBitstringTransformer,
-    BitstringToBitsTransformer,
-    BitstringToUintTransformer,
-    CascadeTransformer,
-    DataTransformer,
-    LookupTableTransformer,
-    UintToBitstringTransformer,
-)
+from core.data_stream import BitsDataStream, DataStream
+from core.data_transformer import DataTransformer
 
 
 class DataCompressor:
@@ -22,7 +14,7 @@ class DataCompressor:
         self.encoder_transform = encoder_transform
         self.decoder_transform = decoder_transform
 
-    def set_encoder_decoder_params(self, data_stream):
+    def set_encoder_decoder_params(self, data_stream: DataStream):
         """
         Usually we will set the self.encoder_transform and self.decoder_transform in this function
         as most of time the parameters of the data_stream are necessary to create the transformers
@@ -30,7 +22,7 @@ class DataCompressor:
         """
         pass
 
-    def encode(self, data_stream):
+    def encode(self, data_stream: DataStream):
         """
         The core encode function of the compressor
         """
@@ -45,53 +37,8 @@ class DataCompressor:
         assert isinstance(output_bits_stream, BitsDataStream)
         return output_bits_stream
 
-    def decode(self, data_stream):
+    def decode(self, data_stream: BitsDataStream):
 
         # input stream to the decoder needs to be a stream of bits
         assert isinstance(data_stream, BitsDataStream)
         return self.decoder_transform.transform(data_stream)
-
-
-class FixedBitwidthCompressor(DataCompressor):
-    """
-    First example of a simple compressor.
-    Encodes each symbol of the data to a fixed bitwidth
-    """
-
-    def set_encoder_decoder_params(self, data_stream):
-        """
-        TODO: add stuff here
-        """
-        # set bit width
-        alphabet = data_stream.get_alphabet()
-        self.bit_width = self._get_bit_width(len(alphabet))
-
-        # set encoder/decoder lookup tables
-        self.encoder_lookup_table = {}
-        self.decoder_lookup_table = {}
-
-        for id, a in enumerate(alphabet):
-            self.encoder_lookup_table[a] = id
-            self.decoder_lookup_table[id] = a
-
-        # create encoder and decoder transforms
-        self.encoder_transform = CascadeTransformer(
-            [
-                LookupTableTransformer(self.encoder_lookup_table),
-                UintToBitstringTransformer(bit_width=self.bit_width),
-                BitstringToBitsTransformer(),
-            ]
-        )
-
-        # create decoder transform
-        self.decoder_transform = CascadeTransformer(
-            [
-                BitsToBitstringTransformer(bit_width=self.bit_width),
-                BitstringToUintTransformer(),
-                LookupTableTransformer(self.decoder_lookup_table),
-            ]
-        )
-
-    @staticmethod
-    def _get_bit_width(alphabet_size) -> int:
-        return int(np.ceil(np.log2(alphabet_size)))
